+++ CotEditor/Sources/FindPanelFieldViewController.swift
@@ -198,6 +198,7 @@ final class FindPanelFieldViewController: NSViewController, NSTextViewDelegate {
         self.applyResult(message: message, textField: self.findResultField!, textView: self.findTextView!)
         
         // dismiss result either client text or find string did change
+        self.removeCurrentTargetObservers()
         self.currentResultMessageTarget = target.layoutManager
         NotificationCenter.default.addObserver(self, selector: #selector(clearNumberOfFound), name: NSTextStorage.didProcessEditingNotification, object: target.textStorage)
         NotificationCenter.default.addObserver(self, selector: #selector(clearNumberOfFound), name: NSWindow.willCloseNotification, object: target.window)
@@ -271,10 +272,7 @@ final class FindPanelFieldViewController: NSViewController, NSTextViewDelegate {
         self.applyResult(message: nil, textField: self.findResultField!, textView: self.findTextView!)
         
         // -> specify the object to remove osberver to avoid removing the windowWillClose notification (via delegate) from find panel itself.
-        if let target = self.currentResultMessageTarget?.firstTextView {
-            NotificationCenter.default.removeObserver(self, name: NSTextStorage.didProcessEditingNotification, object: target.textStorage)
-            NotificationCenter.default.removeObserver(self, name: NSWindow.willCloseNotification, object: target.window)
-        }
+        self.removeCurrentTargetObservers()
     }
     
     
@@ -294,6 +292,18 @@ final class FindPanelFieldViewController: NSViewController, NSTextViewDelegate {
         
         // add extra scroll margin to the right side of the textView, so that entire input can be read
         textView.enclosingScrollView?.contentView.contentInsets.right = textField.frame.width
+    }
+    
+    
+    /// remove observers for result message clear
+    private func removeCurrentTargetObservers() {
+        
+        guard let target = self.currentResultMessageTarget?.firstTextView else { return }
+        
+        NotificationCenter.default.removeObserver(self, name: NSTextStorage.didProcessEditingNotification, object: target.textStorage)
+        NotificationCenter.default.removeObserver(self, name: NSWindow.willCloseNotification, object: target.window)
+        
+        self.currentResultMessageTarget = nil
     }
     
 }