+++ Tests/StringExtensionsTests.swift
@@ -86,6 +86,10 @@ class StringExtensionsTests: XCTestCase {
         XCTAssertEqual(testString.lineNumber(at: 5), 2)
         XCTAssertEqual(testString.lineNumber(at: 6), 3)
         XCTAssertEqual(testString.lineNumber(at: 7), 4)
+        
+        XCTAssertEqual("\u{FEFF}".numberOfLines(in: NSRange(0..<1), includingLastLineEnding: false), 1)  // "\u{FEFF}"
+        XCTAssertEqual("\u{FEFF}\nb".numberOfLines(in: NSRange(0..<3), includingLastLineEnding: false), 2)  // "\u{FEFF}\nb"
+        XCTAssertEqual("a\u{FEFF}\nb".numberOfLines(in: NSRange(1..<4), includingLastLineEnding: false), 2)  // "\u{FEFF}\nb"
     }